GTiff: avoid SetMetadata() to cancel effect of SetGeoTransform() (fixes #6015)
The culprit was that metadata writing calls GetUnitType() which calls
LookForProjection(), which resets the m_bGeoTIFFInfoChanged flag.
Reseting that flag was apparently needed in the past, but is apparently
no longer needed nowadays (there are other resets of it during file opening or
createcopy), at least this doesn't break existing regression tests.