gdalwarp/GDALCreateGenImgProjTransformer2: support identity source geotransform as a valid one (fixes #5954)
Since the origins (2b323d0e164864662dcb3abccf236a20bf391290), the identity
geotransform was considered as not being valid. This was presumably
because at that time some drivers might have returned
GetGeoTransform(...) == CE_None with an identity matrix, whereas there
was none set. There might be some esoteric drivers that still behave
like that, but that should be marginal, and they should rather be fixed.