RPC DEM extraction: use a block base approach to query the DEM
Before this commit, we would query small windows, of growing radius when
querying in the same area, at 'random' locations. When transforming
several points with the same transformer object, there was no guarantee
that the same input coordinates would result in reading the DEM in the
same window. For 'normal' rasters, this didn't make any difference, but
when requesting a VRT with non-integer src/dst offsets or scaling, there
is no guarantee that extracting values from those windows results in
consistent values being returned (can be seen as a bug of the VRT
driver, but it is extraordinary complicated, if not impossible, to offer
such stability property with it).
Consequently let's query the DEM by non-overlapping fixed-size blocks,
which eliminates that issue.