Python bindings: use %feature(shadow) to override Python wrappers (fixes #6304)
When wanting to customize the behaviour of a Python wrapper method, up to now,
we used to override it in a %extend{ %pythoncode %{ %} } section.
The issue is that in the .py file the method was defined twice, with the override
being luckily used, but that is a bit confusing. Using %feature("shadow")
is cleaner in that it only emits our override.