Add OGRLayer::GetGeometryTypes()
This method iterates over features to retrieve their geometry types. This
is mostly useful for layers that report a wkbUnknown geometry type.

By default this method returns an array of nEntryCount entries with each
geometry type (in OGRGeometryTypeCounter::eGeomType) and the corresponding
number of features (in OGRGeometryTypeCounter::nCount).
Features without geometries are reported as eGeomType == wkbNone.