[NTF/nitffile.c] fixed NTF driver bug that prevents adding subsequent TREs after a HEX TRE (fixes #6827)
There was a bug in nitffile.c in method NITFWriteTREsFromOptions. In this method there is a loop. Within this loop, is a string compare method that accepts the length to compare. This string compare method is used to filter out creation_options that do not have the correct prefix (in our case the prefix is TRE=). The string compare length variable is initialized prior to the loop (in our case 4), but once a hex string is passed in, this length is increased by 4. On the next iteration of the loop, this string comparison length is not reset to the original value so all subsequent TREs are filtered out because the string comparison fails (incorrectly).  This commit fixes that issue.