No longer alias VSILFILE* to FILE* in non-DEBUG builds (fixes #6790)
The VSILFILE* type is no longer aliased to FILE* in builds without the DEBUG
define (that is production builds). External code that used FILE* with
GDAL VSI*L API has to be changed to use VSILFILE*.
This aliasing dates back to old times where both types where indifferently
used in the code base. In the mean time, this was cleaned up. But there was a
drawback of having VSILFILE* being either a dedicated type of an alias of
FILE* depending whether DEBUG is defined, especially with the C++ API, for
people building their plugins with DEBUG and running them against a non-DEBUG
GDAL build, or the reverse.