python: fix setup.py dir-list issue on macOS
In finalize_options(), there is a code-path for existing string-style
include/library dirs. In a standalone build this resulted in appending
an empty directory to each list, which led to CC invocations with bare
`-I` and `-L -lgdal`, the latter resulting in modules not being linked
to libgdal, as the second `-lgdal` argument was interpreted as a dir.

Fix by checking whether include_dirs/library_dirs are non-empty before
appending.