Fix integration of find_package2
find_package2 is a custom alternative to regular find_package.
gdal_check_package was meant to not call find_package when
find_pacakge2 returned a positive or negative result. But the
previous guard also interfered with the standard CMake pattern
of repeated calls to find_package.
This commit changes the control over the alternative:
- find_package2 is called exclusively iff a definition exists
  for the module of interest.
- Even this call can be disabled by the user via
  FIND_PACKAGE2_<pk>_ENABLED in order to enable a regular
  find_package call.