autotest: Mark remaining ogr_ngw tests as slow
test_ogr_ngw_test_ogrsf (450 seconds) was marked as slow in
f261a58ea255. Several of the remaining tests  take > 5 seconds to run,
which is longer than other tests that are currently marked as slow
(e.g., test_ogr_wfs_esri_2). But because the faster tests depend on
state set by the slower tests, the entire module needs to be marked as
slow.