Pansharpening: require geotransform on panchromatic and multispectral bands
Pansharpening now requires that panchromatic and multispectral bands have
valid geotransform (in early versions, it was assumed in the case of missing
geotransform that they covered the same geospatial extent).
The undocumented VRT pansharpened MSShiftX and MSShiftY options (and the
corresponding C++ GDALPansharpenOptions::dfMSShiftX and dfMSShiftY members)
have been removed, due to using the inverted convention as one would expect,
and being sub-par solution compared to using geotransform to correlate pixels
of panchromatic and multispectral bands.