GRIB: g2clib: allow negative longitudes in grid templates Lon/Lat, Rotated, Stretched, Stretched & Rotated, LAEA (fixes #7456) and Mercator
It is not obvious that negative longitudes are allowed in the GRIB2
standard. https://old.wmo.int/extranet/pages/prog/www/DPS/FM92-GRIB2-11-2003.pdf
mentions:
"""92.1.8 The longitude values shall be limited to the range 0 to 360 degrees inclusive. Orientation shall be east longitude
positive, with only positive values being used"""

That said wgrib2 in grb2.h reads the center longitude as a signed
integer:
```
```

https://apps.ecmwf.int/codes/grib/format/grib2/templates/3/140 (LAEA) doesn't
specify the signedness, but other templates such as
https://apps.ecmwf.int/codes/grib/format/grib2/templates/3/10 (Mercator)
mentions signed for longitudes...

There is no harm I can think of in allowing more liberarly signed
values for longitudes, latitudes, given the range on how they are
encoded, there is no risk with unsigned values that would overflow the
signed range.

More templates could be edited to allow signed longitudes, but this is
tedious...