docker/ubuntu*/Dockerfile: remove apt lists
Some RUN lines remove the apt lists, and
others do not. Remove the apt lists for
all RUN lines. This shrinks the ubuntu-small-latest-amd64
image size from 402MB to 357MB, and
the ubuntu-full-latest-amd64 image
from 1.86GB to 1.82GB.

As a bonus, this also reduces the differences
between the small and full Dockerfiles.