NITF: fix MIN/MAX_LONG/LAT when reading RPC00B
Last time the computation was modified was in https://github.com/OSGeo/gdal/commit/c654d4fbfb38187c0337a97862d53022ecce6d26
but this appears to be wrong to me as per
http://geotiff.maptools.org/rpc_prop.html, latitudes and longitudes are
normalized to [-1,1] range with the _SCALE/_OFF coefficients.
Consequently, we should do OFF +/- SCALE to get back the initial range.
This analysis is consistent with was is done in the ENVI driver with
```
    CPLsnprintf(sVal, sizeof(sVal), "%.16g",
                CPLAtof(papszFields[3]) - CPLAtof(papszFields[8]));
    SetMetadataItem("MIN_LONG", sVal, "RPC");

    CPLsnprintf(sVal, sizeof(sVal), "%.16g",
                CPLAtof(papszFields[3]) + CPLAtof(papszFields[8]));
    SetMetadataItem("MAX_LONG", sVal, "RPC");

    CPLsnprintf(sVal, sizeof(sVal), "%.16g",
                CPLAtof(papszFields[2]) - CPLAtof(papszFields[7]));
    SetMetadataItem("MIN_LAT", sVal, "RPC");

    CPLsnprintf(sVal, sizeof(sVal), "%.16g",
                CPLAtof(papszFields[2]) + CPLAtof(papszFields[7]));
    SetMetadataItem("MAX_LAT", sVal, "RPC");
```