Revert "Python bindings: define entry_points.console_scripts"
This reverts commit efea24b2b54319ff42e1c623e84e92e64b79b81b.

Cf https://lists.osgeo.org/pipermail/gdal-dev/2023-November/057985.html:
""""
The Python utilities are no longer installed with the .py extension,
this broke their use in QGIS when we removed this in the Debian package
some time ago. Don't know if that's still the case. Either way, this
doesn't seem like an appropriate change for a .1 patch release.

usr/bin/__init__ is now also installed which not correct.
""""