NETCDF: Fix multiple issues with the parser and make it more robust
- vsicurl_streaming
- naked http(s) protocol
- quoted queries
- mismatched quoted paths

I actually tested the API with Open() and all tests are passing.

The antisocial file name with slashed subdatasets is not supported because it
is not compatible with forward paths on mingw64 windows tests, 'NETCDF:a:/blah'
is not a path to a file named 'a' with subdataset '/blah' but a path to a file
named 'blah' on drive 'a', this also on Linux.