VRTComplexSource: fix excessive RAM usage with many sources (fixes #8967, 3.8.0 regression)
m_abyWrkBuffer and m_abyWrkBufferMask were mistakenly put at the
VRTComplexSource level, missing that there can be a big number of
sources whose lifetime is the same as the VRT dataset, and thus it
is inappropriate to have long-lived working buffers at that level.
We can actually use one single instance of them for all sources, so
move that at the dataset level.