/vsicurl/: fix potential multithreaded crash when downloading the same region in parallel and that the download fails
If 2 threads tried to download the same region at the same time, one of
them would wait for the first one to have finished. But if that download
failed, the waiting thread would then wrongly try to unregister the
region in m_oMapRegionInDownload (the first thread has already done
that), resulting in either an assertion in debug mode on
```CPLAssert(oIter != m_oMapRegionInDownload.end())``` in
NotifyStopDownloadRegion() or a crash later when trying to lock
region.oMutex which would be random data.