This commit updates OGC map tiles api to support additional image formats beyond PNG/JPG as well as enables content negotiation.
In our case, the motivating factor is we have map tiles that are stored as GeoTiffs which we would like to share via the OGC Tile API. Tiffs are one of the six predefined encodings allowed in an implementation - see https://docs.ogc.org/is/20-057/20-057.html#toc69.

As part of these changes, collections which do not specify a image type are now supported via content negoation (see https://docs.ogc.org/is/20-057/20-057.html#toc28).

For example, previously this command used to fail and now succeeds.

gdalinfo "OGCAPI:https://test.cubewerx.com/cubewerx/cubeserv/demo/ogcapi/Daraa/collections/AgricultureSrf?f=json"

The reason for failure is the link with a "rel" value of "http://www.opengis.net/def/rel/ogc/1.0/tilesets-map" did not specify a type field. And thus the old code did not recognize it as JPEG or PNG and ignored it.

Finally, this commit removes the hard-coded value of nbands (3 for JPEG and 4 for PNG). Instead, it tries to download a tile and read the actual number of bands. It tries to read tile 0,0,0. Perhaps that is too brittle? But the hard-coded band numbers weren't really correct, at least for JPEG, and definitely not for TIFF.