[openfilegdb] Correctly use "features" as related table type
Previously we reported the table type as "feature", which did
not match the accepted values reported by GDAL_DMD_RELATIONSHIP_RELATED_TABLE_TYPES
for the drivers (and differed from the "features" string used
by other drivers, eg. GPKG).

This is purely a GDAL implementation detail, it does not affect
how the relationships are actually recorded in the GDB (as we
only check for the table type of "media" to apply special
handling)