[gpkg] Also read relationships defined using foreign key constraints
When reading relationships, always read relationships defined using
foreign key constraints regardless of whether or not the related
tables extension is in use.

The related table extension only permits definition of many-to-many
relationships, so there's a strong case for supporting one-to-many
relationships defined outside of this extension. In fact it's what's
recommended upstream: https://github.com/opengeospatial/geopackage/issues/678#issuecomment-1954402113