OpenFileGDB writer: fix corrupted maximum blob size header field in some SetFeature() scenarios (fixes #9388)
When SetFeature() involves generating a row whose blob size is larger
than the maximum one in the file, the file header was no properly
refreshed. This didn't cause issues to the OpenFileGDB driver, which is
robust to that, but ESRI based readers, like ArcMap or the FileGDB SDK,
were unable to read such features.