QuietDeleteForCreateCopy(): forward source dataset open options (#9422)
When using gdal_translate, the source datasource is opened twice. However, the second time does not use any open options specified when orginally opening the data soruce. This is a probelm for OGC Tile data sources if the user specified TILEMATRIXSET or TILEMATRIX because the driver will default to WorldCRS84Quad even when that is not what was originally specified or to a different zoom level.