GTiff + internal libtiff: deal with issues with multi-band PlanarConfig=Contig, LERC and NaN values (fixes #9530)
The LERC codec up-to-now didn't do any special processing when writing a
strile in a multi-band PlanarConfig=Contig file, resulting in
corruptions. Now we generate a mask buffer in that situation. Actually,
if the NaN values are not at the same location accross bands, we need to
write one mask-per-band, which requires liblerc >= 3.0, and in that
situation change the LERC blob to multi-band instead of multi-depth/dim,
with liblerc >= 3.0. When all bands have NaN values at the same
location, we generate a single mask, which is compatible of liblerc <
3.0, but requires at last this version of libtiff to be read.
This is the best solution we come with to fix the issue while having
the maximum backward compatibility.

This also fixes a wrong condition for creation of tiled JPEG where the
condition to pad the right-most and bottom-most tiles was wrongly
written and would have only worked for rasters with same number of tiles
in horizontal and vertical dimension