GDALNoDataMaskBand::IRasterIO(): fix crash on memory allocation failure
The method had a fallback path when allocating a temporary buffer of the
size of the one passed ot the method, that intended to use a less memory
intensive block-based approach, but this was buggy.
Fix this, and add testing for the fallback.
Also, when attempting the fallback, emits a CE_Warning instead of a
CE_Failure.

Fixes https://github.com/rasterio/rasterio/discussions/3028