OpenFileGDB: add partial read-only support for tables with 64-bit ObjectIDs
Tables with non-sparse pages in .gdbtablx are fully supported.

A subset of tables with sparse pages are properly supported (cf
https://github.com/rouault/dump_gdbtable/wiki/FGDB-Spec#trailing-section-16-bytes--variable-number-)
When we detect that we can't properly decode ObjectID in the unsupported
subset of sparse tables, we emit a warning, and turn off use of attribute/spatial indices.

Read-only support for now.