OpenFileGDB writer: .gdbtable header must be rewritten when updating an existing feature at the end of file
The only consequence of the omission of that fix is that the file size
in the .gdbtable could be incorrect. A wrong file size in the header has no
consequence on OpenFileGDB reader/writer. It also doesn't seem to affect
Esri reader side, but it could *potentially* affect the writing side, if
that field was trusted to seek to the end of the file.