Add VSIMultipartUploadXXXXX() API for multi-part upload
Using this API directly is generally not needed, but in very advanced cases,
as VSIFOpenL(..., "wb") + VSIFWriteL(), VSISync(), VSICopyFile() or
VSICopyFileRestartable() may be able to leverage it when needed.

This is only implemented for the /vsis3/, /vsigs/, /vsiaz/, /vsiadls/
and /vsioss/ virtual file systems.

The typical workflow is to do :
 - VSIMultipartUploadStart()
 - VSIMultipartUploadAddPart(): several times
 - VSIMultipartUploadEnd()

If VSIMultipartUploadAbort() is supported by the filesystem (VSIMultipartUploadGetCapabilities()
can be used to determine it), this function should be called to cancel an
upload. This can be needed to avoid extra billing for some cloud storage
providers.