CMake: add infrastructure so we can build plugins that can be loaded against a libgdal that has been without any support for them
Note: this is an ABI break for in-tree drivers, with deferred loading capability,
when built as plugin. That is such a driver built against GDAL 3.9.0
won't be able to be loaded by GDAL 3.9.1.

The CORE_SOURCES option points to a list of files that are used for deferred plugin
loading (cf RFC 96 https://gdal.org/development/rfc/rfc96_deferred_plugin_loading.html),
when the driver is built as a plugin.
Those files are build in libgdal so it has access to plugin metadata and the
identification method, and are typically exported so that the driver can also
access them.
However it has been found that this resulted in the inability of building a
plugin for a libgdal that hadn't been built with any form of support for it,
which is undesirable.
Starting with GDAL 3.9.1, it is strongly recommended to also define NO_SHARED_SYMBOL_WITH_CORE
so that the files pointed by CORE_SOURCES are build twice: once in libgdal
with a "GDAL_core_" prefix, and another time in the plugin itself with a
"GDAL_driver_" prefix, by using the PLUGIN_SYMBOL_NAME() macro of gdal_priv.h

Example in ogrocidrivercore.h:
```
        #define OGROCIDriverIdentify \
           PLUGIN_SYMBOL_NAME(OGROCIDriverIdentify)
        #define OGROCIDriverSetCommonMetadata \
           PLUGIN_SYMBOL_NAME(OGROCIDriverSetCommonMetadata)

        int OGROCIDriverIdentify(GDALOpenInfo *poOpenInfo);

        void OGROCIDriverSetCommonMetadata(GDALDriver *poDriver);
```