GTiff: fix reading angular projection parameters in non-degree unit (typically grads), when reading projection from ProjXXXXGeoKeys
Fixes https://github.com/OSGeo/gdal/issues/10154

libgeotiff is unfortunately inconsistent. When it synthetizes the
projection parameters from the EPSG ProjectedCRS code, it returns
them normalized in degrees. But when it gets them from
ProjCoordTransGeoKey and other Proj....GeoKey's it return them in
a raw way, that is in the units of GeogAngularUnitSizeGeoKey
From GDAL 3.0 to 3.9.0, we didn't take the later case into account, and
lacked a conversion to degree.
And all versions of GDAL <= 3.9.0 when writing those geokeys, wrote
them as degrees, hence this GTIFF_READ_ANGULAR_PARAMS_IN_DEGREE
config option that can be set to YES to avoid that conversion and
assume that the angular parameters have been written as degree.