Arrow/Parquet/generic arrow: add write support for arrow.json extension
Read support was added in 3.9.0 per 21c06b9d486. Now that the arrow.json
canonical extension has been made official, also implement it on the
write side: i.e. in the CreateField() method of the Arrow/Parquet
drivers, and in OGRLayer::GetArrowSchema() when generating a Arrow
schema from a OGR layer definition.