Add read-only raster driver, SNAP_TIFF, for Sentinel Application Processing GeoTIFF files
Such files are formulated in a way that makes it difficult to read them with the
generic GTiff driver:
- single strip of uncompressed Float32 data type (1.7 GB on the sample file I worked on)
- more than 65,535 GCPs in the GeoTIFF TiePoints tag (> 1 million), which make it
  impossible to read with libgeotiff. Furthermore those tiepoints form
  actually a geolocation array, and are thus best represented using the
  GDAL GEOLOCATION formalism
- a unknown TIFF tag of code 65000, which happens to contain XML DIMAP
  metadata of several megabytes. Part of it contains information on the
  nodata value, offset and scale, and thus must be parsed to report
  appropriate band properties

All that combined calls for a specialized driver, instead of cluttering
even more the generic GTiff driver.
To help for that, I've developed a simple TIFF reader in https://github.com/libertiff/libertiff
and included it here.