docker: print HTTP errors on failure
The -q parameter to wget makes it completely
silent, even on faliure which gives build errors
that are difficult to diagnose. Someone
else had the same problem 15 years ago:

https://serverfault.com/questions/70889/wget-to-print-errors-but-nothing-otherwise

and the easiest way to robustly print
errors but nothing else seems to be
to use curl instead.

Since the docker images already include
curl, switch them to use that instead.