Zarr: Check SRS axis order when setting data axis to SRS axis mapping
This fixes a bug that occurred when using gdalwarp on ZARR files that
use EPSG:4326, where the warped output would be lat/lon transposed.

Our understanding of this is that in Zarr, the axes are found by
checking their names and metadata. The existing logic assigns the
HORIZONTAL_X type to the longitude axis. However, OGRSpatialReference
returns an axis mapping of [1, 2] when in authority compliant mode, no
matter whether the CRS is northing/easting or easting/northing oriented.

This change is our attempt to fix the issue without impacting other
formats.

Co-authored-by: Bernhard Stadlbauer <11799671+bstadlbauer@users.noreply.github.com>