ZARR: add support for reading Kerchunk JSON and Parquet reference files.
The driver supports reading "virtual" Zarr datasets where the Zarr logical hierarchy
does not directly translate into a hierarchy of files on the filesystem, but
through a file that contains a store of references to chunk locations (possibly remote).
Such stores are generated by the Python `Kerchunk <https://fsspec.github.io/kerchunk/spec.html>`__
library.

There are 2 types of Kerchunk reference stores:

- JSON reference stores, where the Zarr dataset is entirely described in a single JSON file.
  This comes in two versions: `Version 0 <https://fsspec.github.io/kerchunk/spec.html#version-0>`__
  and `Version 1 <https://fsspec.github.io/kerchunk/spec.html#version-1>`__.

  Such stores can be opened with:

    ::

        gdalmdiminfo "/vsikerchunk_json_ref//path/to/reference.json"

  or sometimes, when the header of the file is sufficient to be recognized
  as a Kerchunk reference store, with just:

    ::

        gdalmdiminfo "/path/to/reference.json"

  Note that the "templates" and "gen" features of Version 1 are not supported.

- Parquet reference stores, where the Zarr dataset is described in a :file:`.zmetadata` file that
  contains only the JSON definitions of the groups and arrays. The chunks themselves
  are cataloged in Parquet files in subdirectories:
  https://fsspec.github.io/kerchunk/spec.html#parquet-references

  Such stores can be opened with:

    ::

        gdalmdiminfo 'ZARR:"/path/to/directory/where/.zmetadata/is/located"'

  That is exactly like a regular Zarr dataset.

  To be able to read pixel values, the :ref:`vector.parquet` driver must be
  available.

As JSON reference stores can be very large and slow to parse, the :oo:`CACHE_KERCHUNK_JSON`
open option can be set to ``YES`` to ask to generate and use a local Parquet
reference store that is cached in :file:`${HOME}/.gdal/zarr_kerchunk_cache`.
This implies the :ref:`vector.parquet` driver is available.

The driver does not rotate cached stores in the local cache. It is the responsibility
of the user to manage its content and remove obsolete datasets.

It is also possible to convert JSON reference store into a Parquet one using
the :co:`CONVERT_TO_KERCHUNK_PARQUET_REFERENCE` creation option set to ``YES``.

For example:

::

    gdal_translate -of ZARR -co CONVERT_TO_KERCHUNK_PARQUET_REFERENCE=YES store.json store.parq