GDALOverviews: Limit external file size in GDALRegenerateOverviewsMultiBand (#12392)
## What does this PR do?
In GDALRegenerateOverviewsMultiBand() (ROMB), when too much memory would be needed for generating an overview block, the overviews are generated in a temporary file and then copied to the actual overview dataset. The size of this temporary file is not limited, and it has extra problems with pixel interleaved output, since the bands are copied one at a time for the full overview.
This issue was reported in issue #12303 and was partially addressed in #12307, which raised the allowable memory limit which triggers the problem code path, but did not limit the size of the temporary file or the pixel interleaved problems.

This PR addresses the remaining problems by chunking the output overview when possible, and using the temporary file path for a single chunk at a time.

## What are related issues/pull requests?
#12303, #12307

## Test case
To trigger the code path in question, set GDAL_OVER_CHUNK_MAX_SIZE_FOR_TEMP_FILE to 10485760 and build power of 2 overviews for a 11000 by 11000 8bit, pixel interleaved RGB MRF with a blocksize of 1024

```
gdal_translate -outsize 11000 11000  -co BLOCKSIZE=1024 any_small_rgb.jpg test_file.mrf
gdaladdo -r average test_file.mrf
```

## Changes

The "chunking" code triggers when a temporary file would be used, but an output chunk (tile) smaller than the whole overview does exist. The chunk would still exceed the memory usage limit. It then calls ROMB() for every chunk, with extra parameters that limit the processing extent.
On the next pass through ROMB, the external file will be created, this time limited to the chunk extent. This part was also modified, using a VRT to effect the scaling, with the correct resampling, before writing the chunk data in the external file.
Finally, the chunk data is transferred from the temporary output to the destination overview.

In addtion, an extra flag had to be added to the RasterIO ExtraArguments structure, to indicate that the source overviews should not be used for the respective operation, to avoid having the input data read from the same overview that is being generated.

---------

Co-authored-by: Even Rouault <even.rouault@spatialys.com>