TileDB: add support for /vsiaz/ (#12666)
It also adds a check if the group/array exists in Open as there isn't a way to prevent a driver scan when returning false from Identify.

An example is if a user passes in that allowed drivers are TileDB to OpenEx (in read mode) and hence Identify returns true but the actual group/array does not exist. If Identify returns false then a driver scan continues.