MIraMonRaster: fixing fuzzer Issue 443651667
A band name may be empty only if it is the only band present in the REL file. Otherwise, inferring the band name from the REL filename is considered an error.  Consequently, for a REL file containing exactly one band, if the band name is empty, it shall be inferred from the REL filename.
Example: REL: testI.rel  -->  IMG: test.img