Create distribution CI workflows
[skip ci]

On a git-tag starting with "v", versioned docker images are created with
attestation, and with both amd64 and arm64 platforms and "joined" via
a manifest.

Furthermore mkgdaldist.sh is run, creating the source release artifacts,
which are uploaded to a draft release on github.  The source releases
are also have attestation run on them.