Revise link lib flattening
Do not flatten a target more than once. In particular abseil (via arrow/
parquet) comes with a complex dependency network, and cmake configuration
failed in some situations.

The new implementations starts the flattening from the end of the of the
link lib list. In this way, link order requirements for traditional linkers
can be satisfied despite targets being flattened only on first visit.