[rasterize] Fix crash when empty bbox
Fix #13498 by checking for empty bbox, but
also implements deeper checks at the library
level to prevent crashes when the geotransformer
is reset to null after the use of overviews
triggers a failure to compute the base transformer.

After the bbox input check I couldn't find a way to
trigger the code path that causes the crash in the first
place, so the library fixes were manually tested before
the input check was applied.