BMP: allow reduced bpp on conversion
Currently the bits per pixel in a GeoTIFF created from a BMP is at least
8.  For palette indexed images with a small number of colours this means
the output file is much larger than it needs to be.  In particular it is
much larger than the BMP is first converted to a GIF or PNG externally.
Fix this by setting NBITS in the metadata. The GeoTIFF file from direct
conversion of a BMP file is now (sha256sum) identical to that from a PNG
file converted from the BMP file.