BMP: fix RLE4 encoded mode
There was a subtle bug in the RLE4 decoder for encoded mode alternating
value runs. In the first pixel value of the alternating run depended on
whether the length of the run is odd or even. So `05 12` was decoded to
`1 2 1 2 1` (correct) and `06 12` was (incorrectly) decoded to `2 1 2 1 2 1`.
This change almost matches the decoding of absolute mode a few lines
below.

This bug was found and the fix verified by comparing a GeoTIFF created
by gdal_translate directly from a RLE4 compressed BMP with a
GeoTIFF created by gdal_translate from a PNG converted with ImageMagick
from the RLE4 BMP file.

A test file in RLE4 format has been added generated from 4bit_pal.bmp.
This test verifies the fix in this commit.